Update patch set 1
Patch Set 1: No score

StaticLayout has no way to change the text once set (in fact, Layout has no way, which means that none of the layouts will). This means that solving the core of the problem will require some restructuring of the entire CalendarView, to store a layout for each text field.

Such a change goes somewhat deeper than I intended to go. It is, in all likelihood, a necessary thing for reasons unrelated to BiDi.

I'll see if I have the time to work on it, but I would really love to hear from whoever it is that is currently in charge of Calendar (Michael Chan?)

Shachar

Patch-set: 1
Label: Code-Review=0