Update patch set 1
Patch Set 1:

Hi James,

Thank you for submitting the patch. The start day of the week change is a frequently requested feature and it is on my todo list. You got to it before I did. I will take care of integrating it into the master branch with some edits e.g. changing tabs to spaces and adding "Saturday" and "Use default" which is locale specific. "Use default" is needed so that we can provide a good default for the user and preserve the current behavior.

As for UI changes, the UI designer has a new ideas although it hasn't been finalized yet.  I will hold off on integrating your change to master for now.

Thanks,
Mike

Patch-set: 1