Update patch set 8
Patch Set 8: Fails

This conflicts with the current master tree (which got a lot of changes in the last few week as froyo was getting open-sourced). It appears that a lot of this change already exists in the master tree, though, and that there's very little left to do.

If this change is still relevant, please rebase it on the latest master and upload a replacement.

Patch-set: 8
Reviewer: Gerrit User 1000660 <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=-1