Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

I think it is frustrating enough to try and contribute to the Android project without having to adhere to policies one cannot see, which directly contradict the publicly published policies.

This patch, almost exactly the same way as it is today, was first submitted on Dec 23, 2009, in change #12661. Ever since, it has been ignored, amended for white spaces, ignored again, dropped, waited for me to be able to test it again. In all of that time it has undergone no significant change whats o' ever. I think it would be cruel to reject it now because of this reason.

Patch-set: 2