Update patch set 2
Patch Set 2: (1 inline comment)

Regarding the code style guidelines - I am well aware of it. It does not talk about 80 characters per line, only of 100. RomainGuy was referring to a rule that does not exist there, which is why I called it unfair.

Regarding the StaticLayout comment - I actually do suspect your framework. It behaves fine on my device running Froyo with the code as is in the patch. Also, my understanding of the docs and code, that is precisely what TruncateAt::END should do.

Shachar

Patch-set: 2