Update patch set 1
Patch Set 1:

You really want someone at Google to look at this.

From my point of view I dont like this kind of changes very much as it does not add alot of value but changes alot of files, and thus adding to the cost of maintaining external branches.

Patch-set: 1