Update patch set 2
Patch Set 2:

Thanks for the CL. The code base was heavily modified/refactored in Honeycomb.  CalendarView is now DayView. I will manually create a change to cover that change in Master.  The changes in the other two files do not apply anymore.  That should be the last place with this problem.

This CL should be safe for the GB branch but should not be merged internally.

Patch-set: 2