Update patch set 4
Patch Set 4:

We were unable to auto-submit your change due to the following error: Failed to submit 3 changes due to the following problems:

Change 1752401: Depends on change that was not submitted. Commit 2690d7f95af434733bdfab840761daeeaaa519b7 depends on commit 226b097d058840e6474f549b4b433560d9b5cdde which cannot be merged. Is the change of this commit not visible to 'treehugger-gerrit@google.com' or was it deleted?

Change 1752402: Depends on change that was not submitted. Commit 7b5ca1506c132f23ff36113313f22c6668772ecf depends on commit 226b097d058840e6474f549b4b433560d9b5cdde which cannot be merged. Is the change of this commit not visible to 'treehugger-gerrit@google.com' or was it deleted?

Change 1752403: Depends on change that was not submitted. Commit d0bfa82e4fe8fe371b8086989093295df070a688 depends on commit 226b097d058840e6474f549b4b433560d9b5cdde which cannot be merged. Is the change of this commit not visible to 'treehugger-gerrit@google.com' or was it deleted?



 view details: https://android-build.googleplex.com/presubmit-status?change_id=1752401&revision_id=4&host=android&id=4565226607509504

Patch-set: 4
Tag: autogenerated:TreeHugger:Autosubmit-Fail