Revert "CalendarTests minsdk  bump from 19 to 21"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b84f2962-13de-4da3-b756-298b8f567a31). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 348584791
Change-Id: I57367aa9bd4de94d1c96fa0e185f6d4fdfd95c36
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com